tests: cloudflare changed status code, breaking some tests
CloudFlare has broken API compatibility by changing the error status
codes returned for some cases. This broke some tests on the library.

Since this library exposes the original status code, users of the
library are recommended to check code that uses them against the
CloudFlare documentation, to look for changes.

Affected cases on this library:

1. When an "A" (IPv4 address) record is being created, if it already
   exists the API used to produce error code 81057. Now it returns code
   81058.
2. When an "A" DNS record is being created, but the content is not a
   valid IPv4 address, the response code used to be CF error "1004: DNS
   Validation Error". Now it is "9005: Content for A record must be a
   valid IPv4 address."